+++ src/pydiverse/transform/__init__.py
@@ -1,4 +1,5 @@
-from __future__ import annotations
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
 from ._internal.pipe.pipeable import verb
 from ._internal.pipe.table import Table, backend, is_sql_backed
@@ -7,9 +8,10 @@ from .extended import *
 from .extended import __all__ as __extended
 from .types import *
 from .types import __all__ as __types
+from .version import __version__
 
 __all__ = (
-    ["Table", "ColExpr", "Col", "verb", "backend", "is_sql_backed"]
+    ["__version__", "Table", "ColExpr", "Col", "verb", "backend", "is_sql_backed"]
     + __extended
     + __types
 )