+++ src/pydiverse/transform/_internal/backend/impl_store.py
@@ -1,4 +1,5 @@
-from __future__ import annotations
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
 import dataclasses
 import inspect
@@ -14,7 +15,7 @@ class ImplStore:
 
     impl_trie: dict[Operator, SignatureTrie]
     default_impl: dict[Operator, Callable | None]
-    impl_manager: ImplContextManager
+    impl_manager: "ImplContextManager"
 
     def __init__(self) -> None:
         self.impl_trie = dict()