+++ src/pydiverse/transform/_internal/backend/targets.py
@@ -1,8 +1,9 @@
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
+
 # This module defines the config classes provided to the user to configure
 # the backend on import / export.
 
-
-from __future__ import annotations
 
 import sqlalchemy as sqa