+++ src/pydiverse/transform/_internal/pipe/functions.py
@@ -1,6 +1,7 @@
-# ruff: noqa: A002
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
-from __future__ import annotations
+# ruff: noqa: A002
 
 from collections.abc import Iterable
 from typing import Any, overload