+++ src/pydiverse/transform/_internal/util/__init__.py
@@ -1,3 +1,8 @@
-from __future__ import annotations
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
 from .reraise import reraise
+
+__all__ = [
+    "reraise",
+]