+++ tests/util/__init__.py
@@ -1,3 +1,9 @@
-from __future__ import annotations
+# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# SPDX-License-Identifier: BSD-3-Clause
 
 from .assertion import assert_equal, assert_result_equal
+
+__all__ = [
+    "assert_equal",
+    "assert_result_equal",
+]